Hashing the transient key to prevent key collisions when long keys are used.  Since we're using an MD5 hash for the key, we know the POSTed "key" value in TLC_Transient_Update_Server::init() will be 32 characters long, so this lets us do additional validation to make sure we're getting a valid request before trying to cache it.
More changes to TLC_Transient_Update_Server::init() -- checking ( isset( $update[0] ) ) instead of just ( $update ) to verify that $update != false and $update is an array at the same time, before trying to interact with it.  Also verifying the information stored in the array is what's expected.  Calling tlc_transient( $key, 'nohash' ) when setting the transient data to prevent double-hashing the key.